FRN-2: Little fixes to furniture configuration
