+++ src/main/kotlin/ru/sinitsynme/furnituretasktracker/furniture/model/FurnitureModelDto.kt
@@ -8,3 +8,8 @@ data class FurnitureModelResponseDto(
     val modelId: Long,
     val name: String
 )
+
+fun FurnitureModel.toResponse() = FurnitureModelResponseDto(
+    modelId = modelId!!,
+    name = name
+)
